revenua

+++ lib/data/repositories/orders/orders_repository.dart
@@ -1,6 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/foundation.dart';
 import 'package:get/get.dart';
+import 'package:intl/intl.dart';
 import 'package:sports_shoe_store/data/repositories/authentication/authentication_repository.dart';
 import 'package:sports_shoe_store/features/shop/models/order_model.dart';
 import 'package:sports_shoe_store/utils/constants/enums.dart';
@@ -40,17 +41,35 @@ class OrderRepository extends GetxController{
     }
   }
 
+  Stream<List<OrderModel>> getOrderStream() {
+    return _db.collection('Orders').snapshots().map((snapshot) {
+      return snapshot.docs.map((doc) => OrderModel.fromQuerySnapshot(doc)).toList();
+    });
+  }
+
   ///xoa
 
-  Future<void> deleteOrder(String userId, String orderId) async {
+  Future<void> deleteOrder(String userId ,String orderId,) async {
     try {
-      // Delete order from the user's subcollection
+      final orderDoc = await _db.collection('Users').doc(userId).collection('Orders').doc(orderId).get();
+      if (!orderDoc.exists) {
+        throw 'Order does not exist in user\'s collection.';
+      }
+
+      // Xóa đơn hàng từ subcollection của người dùng
       await _db.collection('Users').doc(userId).collection('Orders').doc(orderId).delete();
-      // Optionally, delete order from the main orders collection if it exists there
-      await _db.collection('Orders').doc(orderId).delete();
+
+      // Xóa đơn hàng từ collection chính nếu nó tồn tại ở đó
+      final mainOrderDoc = await _db.collection('Orders').doc(orderId).get();
+      if (mainOrderDoc.exists) {
+        await _db.collection('Orders').doc(orderId).delete();
+      }
+
+      print('Order successfully deleted');
     } catch (e) {
       print('Error deleting order: $e');
-      throw 'Failed to delete order';
+      // Thông báo lỗi chi tiết cho người dùng
+      throw 'Failed to delete order: $e';
     }
   }
 
@@ -74,15 +93,8 @@ class OrderRepository extends GetxController{
       throw 'Something went wrong while fetching order information. Try again later';
     }
   }
-  // New method to fetch all orders for admin
-  Future<List<OrderModel>> fetchAllOrders() async {
 
-      // Lấy dữ liệu trực tiếp từ server
-      final result = await _db.collectionGroup('Orders').get();
-      return result.docs.map((e) => OrderModel.fromSnapshot(e)).toList();
-
-  }
-
+// New method to fetch all orders for admin
 
   Future<List<OrderModel>> fetchUserOrdersAdmin(String userId) async {
     try {
@@ -95,6 +107,31 @@ class OrderRepository extends GetxController{
     }
   }
 
+  Future<List<OrderModel>> fetchUserOrdersAdmin2(String userId, {String? orderId}) async {
+    try {
+      if (userId.isEmpty) {
+        throw 'Unable to find user information. Try again in a few minutes.';
+      }
+
+      // Tham chiếu đến bộ sưu tập đơn hàng của người dùng
+      final ordersCollection = _db.collection('Users').doc(userId).collection('Orders');
+
+      // Kiểm tra nếu orderId được cung cấp
+      final QuerySnapshot result;
+      if (orderId != null) {
+        // Lấy đơn hàng cụ thể dựa trên orderId
+        result = await ordersCollection.where('id', isEqualTo: orderId).get();
+      } else {
+        // Lấy tất cả các đơn hàng nếu không có orderId
+        result = await ordersCollection.get();
+      }
+
+      // Chuyển đổi kết quả thành danh sách các OrderModel
+      return result.docs.map((e) => OrderModel.fromSnapshot(e)).toList();
+    } catch (e) {
+      throw 'Something went wrong while fetching order information. Try again later';
+    }
+  }
 
 
   //save and update orders
@@ -123,24 +160,23 @@ class OrderRepository extends GetxController{
     try {
       List<OrderModel> receivedOrders = [];
 
-      // Nếu có các đơn đặt hàng tạm thời, thêm chúng vào danh sách
-      if (temporaryOrders.isNotEmpty) {
-        receivedOrders.addAll(temporaryOrders.where((order) => order.status == OrderStatus.received));
-      }
+      // Lấy ngày hôm nay
+      DateTime now = DateTime.now();
+      DateTime today = DateTime(now.year, now.month, now.day);
 
       // Lấy tất cả các tài liệu người dùng
-      final usersSnapshot = await _db.collection('Users').get();
+      final usersSnapshot = await FirebaseFirestore.instance.collection('Users').get();
 
-      // Lặp qua từng tài liệu người dùng để tìm đơn hàng có trạng thái "Received"
+      // Lặp qua từng tài liệu người dùng để tìm đơn hàng có trạng thái "Received" và searchDate là ngày hôm nay
       for (var userDoc in usersSnapshot.docs) {
-        final ordersSnapshot = await _db
+        final ordersSnapshot = await FirebaseFirestore.instance
             .collection('Users')
             .doc(userDoc.id)
             .collection('Orders')
             .where('status', isEqualTo: 'OrderStatus.received')
             .get();
 
-        // Thêm các đơn hàng có trạng thái "Received" vào danh sách
+        // Thêm các đơn hàng có trạng thái "Received" và searchDate là ngày hôm nay vào danh sách
         receivedOrders.addAll(ordersSnapshot.docs.map((e) => OrderModel.fromSnapshot(e)).toList());
       }
 
@@ -150,6 +186,7 @@ class OrderRepository extends GetxController{
       throw 'Something went wrong while fetching received orders. Try again later';
     }
   }
+
 
   // Phương thức để thêm đơn đặt hàng vào biến tạm thời
   void addTemporaryOrders(List<OrderModel> orders) {
